requirements.txt
# tensorflow>=2.7.0  # Required to run the model
python-dotenv>=0.19.2
loguru>=0.5.3
tqdm>=4.62.3
# ipyplot>=1.1.0
numpy>=1.21.5
ray>=1.11.0
GPUtil>=1.4.0
requests>=2.27.1
